refactor: add grpc-practice
