+++ spring-java/grpc-practice/app/src/main/proto/DemoTest.proto
@@ -1,0 +1,20 @@
+syntax = "proto3";
+
+// specify package for generated java classes
+option java_package = "grpc.practice";
+
+service DemoTest {
+  rpc Unary(DemoRequest) returns (DemoResponse) {}
+  rpc ServerSideStreaming(DemoRequest) returns (stream DemoResponse) {}
+  rpc ClientSideSteraming(stream DemoRequest) returns (DemoResponse) {}
+  rpc BidirectionalStreaming(stream DemoRequest) returns (stream DemoResponse) {}
+}
+
+message DemoRequest {
+  int32 a = 1;
+  int32 b = 2;
+}
+
+message DemoResponse {
+  int32 result = 1;
+}
